# app/generated/schema/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Spark Studio Salem ACCX Open Access 4.0 API",
    "description": "This service provides a method for interacting with the ACCX Open Access 4.0 module.",
    "version": "1.0",
    "contact": {}
  },
  "paths": {
...
    "/card/{cardPosition}": {
      "delete": {
        "operationId": "cardDelete",
        "x-eov-operation-handler": "cardDelete",
        "tags": [
          "Cards"
        ],
        "parameters": [
          {
            "name": "cardPosition",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CardPosition"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "cardGet",
        "x-eov-operation-handler": "cardGet",
        "tags": [
          "Cards"
        ],
        "parameters": [
          {
            "name": "cardPosition",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CardPosition"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "cardPut",
        "x-eov-operation-handler": "cardPut",
        "tags": [
          "Cards"
        ],
        "description": "Update an existing card slot, changing its permissions group or badge number. In order to prevent accidentally overwriting an existing card entry the current badge value must also be submitted as part of this request.",
        "parameters": [
          {
            "name": "cardPosition",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CardPosition"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "badge": {
                    "$ref": "#/components/schemas/Badge"
                  },
                  "currentBadge": {
                    "$ref": "#/components/schemas/Badge"
                  },
                  "permissions": {
                    "$ref": "#/components/schemas/Permissions"
                  }
                },
                "required": [
                  "badge",
                  "currentBadge",
                  "permissions"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "string"
      },
      "Badge": {
        "type": "string"
      },
      "Card": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/CardPosition"
          },
          "badge": {
            "$ref": "#/components/schemas/Badge"
          },
          "permissions": {
            "$ref": "#/components/schemas/Permissions"
          }
        },
        "required": [
          "id",
          "permissions",
          "badge"
        ]
      },
      "CardPosition": {
        "type": "string"
      },
      "Permissions": {
        "type": "integer",
        "minimum": 0,
        "maximum": 255,
        "description": "A Permission is a byte value between 0 and 255 that can be used to group cards into groups of behavior."
      }
    }
  }
}